requirements.txt
# Core data analysis
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.0.0  # Required for reading Excel files (.xlsx) with pandas

# Statistical modeling
statsmodels>=0.14.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0

# Logs/debugging
loguru

# linear models
linearmodels

# visual outputs
seaborn

# streamlit
streamlit
streamlit-flow-component # used for decision tree
plotly>=5.0.0  # for interactive 3D visualizations
